requirements.txt
streamlit
streamlit-chat
python-dotenv
transformers
torch
langchain
langdetect
faiss-cpu
edge_tts
SpeechRecognition
chromadb
pyaudio
pypdf
